Rewrite in ES2015
Versioned as 1.0.0-alpha.1 -- the alpha is important.

Functionally, there is (or should be) no change, but this is going to be
the first step toward a fairly significant rewrite. Tests are (finally)
coming, it's in the (hopefully) more accessible ES2015, the code will be
a little more modular, the plugin will have a friendlier API, etc.

Rewriting in ES2015 also involved a significant overhaul of the build,
so I took this opportunity to split out the IntervalTree into my other
library of data structures and algorithms. Now, using Webpack, it can
be imported and *only* the IntervalTree part of the code will actually
be merged into the final built file.

Also tossed in some Tern-related files, for anyone using that. You should,
it's neat.